requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
openai==2.15.0
python-dotenv==1.2.1
pydantic==2.12.5
httpx==0.28.1
